Assume nitrokey 0.4 compatible interface
Nitrokey 0.4 provides a major revamp of the public APIs. One of those
changes is the introduction of a Manager object that ensures that only a
single device connection can exist at any point in time (a constraint
ultimately imposed by libnitrokey).
This change adjusts the crate to be compatible with nitrokey 0.4.