Update nitrokey dev-dependency to 0.4.0-alpha.3
When releasing version 0.3.0 of the crate we had used the version
constraint >=0.4.0-alpha.2 for the dev-dependency to nitrokey. That was
done intentionally, as cargo did not allow us to release a version of
this crate that depended (only) on an unpublished version of another
crate. That's a problem because nitrokey has a dependency on this crate,
causing a catch-22 situation.
This change corrects this constraint now that nitrokey-0.4.0-alpha.3 has
been released.
Looking into this issue some more, it should be possible to just publish
a package that still have a local path override inside of Cargo.toml
(which has to be used for testing anyway). Cargo will ensure to strip
that off before upload.