Use nitrokey::force_take over nitrokey::take in tests
With this change we use nitrokey::force_take instead of nitrokey::take
in the tests. The difference between the two functions is that the
former ignores poison errors on the underlying mutex, which are emitted
when code panics while a mutex is held. However, tests generally are
isolated from each other (as that is one of the prerequisites for
running them in parallel) and so the purpose of those poison errors
does not really apply here.