Remove SupportedDevice::None variant
In the past we have supplied a connected device to the test functions.
However, with version 0.4 of the nitrokey crate and the global manager
approach used to prevent concurrent accesses, that is not necessarily
always sufficient: in the nitrocli crate we have no way to pass in a
manager object to use but use the global one directly. Doing so will
always lead to a clash because we cannot acquire two references to it at
the same time.
To alleviate this problem, the plan is to not only be able to pass a
connected device to a test function, but also merely a model. In such a
scenario, the "framework" would not have to acquire a reference to the
manager before running the test but could leave that to the test itself,
preventing the aforementioned problem.
As a first step moving us into this direction, this change removes the
SupportedDevice::None variant and replaces it with usage of an
Option<SupportedDevice> instead.