Introduce ArgumentType enum
In order to support passing a Model argument to a test function and not
just a ready-connected Device object we need to restructure some of the
logic for emitting code. In particular, we eventually want to be able to
tell the existing expand_wrapper function (as well as expand_arg &
expand_call) to accept the type of argument for which we want to emit a
function.
This change introduces a new enum, ArgumentType, that will be used for
this purpose in the future.