Remove EmittedDevice enum
This patch removes the EmittedDevice from the code base. This type was
introduced to describe what kind of Nitrokey to emit a function for, but
it complicates the introduction of a Model argument for test functions.
So with this change we replace it with usage of a SupportedDevice
object describing the device to emit code for and the previously
introduced ArgumentType describing the type of argument.