Introduce device filtering support
With the upcoming support for passing a Model to a user's test
function we have a problem: unlike devices, for which there exist
dedicated type representing the different Nitrokey variants (Pro &
Storage), the Model is just an enum that cannot tell us anything about a
desired variant. Hence, we need a different way to allow the user to
specify the Nitrokey device model on which a test function should be
invoked.
This patch introduces such functionality in the form of a device filter.
A filter is specified as part of the test annotation, like so:
  #[nitrokey_test::test(pro)]

Such a filter will typically restrict the devices a test is invoked
with. For example, a DeviceWrapper argument can now be constraint to
actually only ever take on the Pro (or Storage) variant. As a special
case, when no argument is accepted by a test function but a filter is
provided, the function will only run when the filtered-for device is
present.