Add support for nitrokey::Model argument
This change finally introduced support for accepting a nitrokey::Model
argument to all test case functions. Providing a Model as opposed to a
connected Device allows the test function to acquire a reference to the
global manager object used by nitrokey 0.4, which otherwise is held by
the test function's caller (i.e., code emitted by this crate). This is a
necessary feature for the nitrocli program, which has no way to pass in
a manager reference to the code under test.