Add Librem Key support
This change adds supported for the Librem Key. Testing on this device
works similar to testing on the Nitrokey Pro or Storage in that it can
be specified in various ways and will implicitly be provided when a test
accepts a "generic" device argument. Furthermore, we now have a another
NITROKEY_TEST_GROUP variant, librem.